Refactor `JsonObject` type, move it to "Data.Json.Util".
