+++ libs/wire-api/src/Wire/API/Routes/Public/Cannon.hs
@@ -17,11 +17,8 @@
 
 module Wire.API.Routes.Public.Cannon where
 
-import Data.Aeson qualified as A
 import Data.Id
-import Data.OpenApi qualified as O
-import Data.Schema as S
-import Imports
+import Data.Json.Util
 import Servant
 import Servant.API.Extended
 import Wire.API.Routes.API
@@ -52,20 +49,6 @@ type CannonAPI =
                :> "config-options-cannon"
                :> Get '[JSON, YAML] JsonObject
            )
-
--- | Arbitrary aeson object value with helpful {to,from}json instances and schema.
-newtype JsonObject = JsonObject {unJsonObject :: A.Object}
-  deriving newtype (Eq, Ord, Show)
-  deriving (O.ToSchema) via (Schema JsonObject)
-
-instance A.FromJSON JsonObject where
-  parseJSON = A.withObject "Object" (pure . JsonObject)
-
-instance A.ToJSON JsonObject where
-  toJSON (JsonObject obj) = A.Object obj
-
-instance S.ToSchema JsonObject where
-  schema = named "Object" $ unJsonObject .= (JsonObject <$> S.jsonObject)
 
 data CannonAPITag
